Session Class and Default Constructors
The session class has been created and includes arrays of actuators. LED and Actuator classes cannot be in the same array so have been separated.

I have added default constructors for LED and Buzzer classes as it was causing errors. Sensors are next then we can add some actual functionality.